@@ -3,25 +3,30 @@ using System.Linq;
 using System.Threading.Tasks;
 using Azure_AD_Users_Publisher.Services.Models;
 using Azure_AD_Users_Shared.Models;
+using LazyCache;
 using Microsoft.Extensions.Logging;
 
 namespace Azure_AD_Users_Publisher.Services
 {
     public class SalesforceMessageProcessor : IMessageProcessor
     {
+        private const string _cacheKeyPrefix = "_SalesforceMessageProcessor_";
+
+        private readonly IAppCache _cache;
         private readonly ILogger<SalesforceMessageProcessor> _logger;
         private readonly IHISCTokenService _tokenService;
         private readonly IProgramDataService _programDataService;
         private readonly ITimeZoneService _timeZoneService;
-        private readonly ISalesforceUserPublishService _salesforceUserPublishService;
+        private readonly ISalesforceUserService _salesforceUserService;
 
-        public SalesforceMessageProcessor(ILogger<SalesforceMessageProcessor> logger, IHISCTokenService tokenService, IProgramDataService programDataService, ITimeZoneService timeZoneService, ISalesforceUserPublishService salesforceUserPublishService)
+        public SalesforceMessageProcessor(IAppCache cache, ILogger<SalesforceMessageProcessor> logger, IHISCTokenService tokenService, IProgramDataService programDataService, ITimeZoneService timeZoneService, ISalesforceUserService salesforceUserService)
         {
+            _cache = cache;
             _logger = logger;
             _tokenService = tokenService;
             _programDataService = programDataService;
             _timeZoneService = timeZoneService;
-            _salesforceUserPublishService = salesforceUserPublishService;
+            _salesforceUserService = salesforceUserService;
         }
 
         public async Task ProcessUser(AzureActiveDirectoryUser user)
@@ -30,10 +35,10 @@ namespace Azure_AD_Users_Publisher.Services
             var syncUserToSalesforce = await CheckUserFranchiseAgainstFranchiseSource(user, ProgramDataSources.Salesforce, true, false);
             if (syncUserToSalesforce)
             {
-                if (user.DeactivationDateTimeOffset.HasValue)
+                if (user.DeactivationDateTimeOffset.HasValue && await UserExistsAndIsActiveInSalesforce(user))
                 {
                     _logger.LogInformation($"User will be Deactivated: {json}");
-                    await _salesforceUserPublishService.Deactivate(user.ExternalId);
+                    await _salesforceUserService.Deactivate(user.ExternalId);
                 }
                 else
                 {
@@ -62,15 +67,25 @@ namespace Azure_AD_Users_Publisher.Services
                     };
 
                     _logger.LogInformation($"User will be Published: {json}");
-                    await _salesforceUserPublishService.Publish(salesforceUser);
+                    await _salesforceUserService.Publish(salesforceUser);
                 }
 
-                _logger.LogInformation($"Publish Count: {_salesforceUserPublishService.PublishCount}, Deactivation Count: {_salesforceUserPublishService.DeactivationCount}, Error Count: {_salesforceUserPublishService.ErrorCount}");
+                _logger.LogInformation($"Publish Count: {_salesforceUserService.PublishCount}, Deactivation Count: {_salesforceUserService.DeactivationCount}, Error Count: {_salesforceUserService.ErrorCount}");
             }
             else
             {
                 _logger.LogInformation($"User will NOT be sent to Salesforce: {json}");
             }
+        }
+
+        private async Task<bool> UserExistsAndIsActiveInSalesforce(AzureActiveDirectoryUser user)
+        {
+            // note: using the default caching duration of 20 minutes
+            var allUsers = await _cache.GetOrAddAsync($"{_cacheKeyPrefix}AllUsers", _salesforceUserService.RetrieveAllUsers);
+            var isActive = allUsers.records.Any(sfUser => sfUser.IsActive 
+                                                  && sfUser.HI_GUID__c != null 
+                                                  && sfUser.HI_GUID__c.Equals(user.ExternalId));
+            return isActive;
         }
 
         private async Task<T> CheckUserFranchiseAgainstFranchiseSource<T>(AzureActiveDirectoryUser user, ProgramDataSources source, T success, T fail)
