@@ -1,4 +1,5 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
 using Azure_AD_Users_Publisher.Services;
@@ -18,12 +19,14 @@ namespace Azure_AD_Users_Publisher.Controllers
         private readonly IHISCTokenService _tokenService;
         private readonly IProgramDataService _programDataService;
         private readonly ITimeZoneService _timeZoneService;
+        private readonly IMessageProcessor _messageProcessor;
 
-        public PublishController(IHISCTokenService tokenService, IProgramDataService programDataService, ITimeZoneService timeZoneService)
+        public PublishController(IHISCTokenService tokenService, IProgramDataService programDataService, ITimeZoneService timeZoneService, IMessageProcessor messageProcessor)
         {
             _tokenService = tokenService;
             _programDataService = programDataService;
             _timeZoneService = timeZoneService;
+            _messageProcessor = messageProcessor;
         }
 
         [ProducesResponseType(typeof(FranchiseResults), StatusCodes.Status200OK)]
@@ -66,6 +69,26 @@ namespace Azure_AD_Users_Publisher.Controllers
             
             return Ok(user);
         }
+
+        // note: quick way to test a given user, not needed for production, still determining if we need this for testing
+        //[ProducesResponseType(StatusCodes.Status200OK)]
+        //[ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        //[Produces("application/json")]
+        //[HttpGet("salesforceUser")]
+        //public async Task<IActionResult> IsSalesforceUser()
+        //{
+        //    var user = new AzureActiveDirectoryUser
+        //    {
+        //        // db12fc65-c9e1-4c32-a9b8-97b14229d166 -- IsActive: false
+        //        // e62d778a-db0a-45d4-8829-e638065ae248 -- IsActive: true
+        //        ExternalId = "db12fc65-c9e1-4c32-a9b8-97b14229d166",
+        //        DeactivationDateTimeOffset = DateTimeOffset.Now.AddDays(-1)
+        //    };
+
+        //    await _messageProcessor.ProcessUser(user);
+            
+        //    return Ok();
+        //}
 
         [ExcludeFromCodeCoverage]
         public class FranchiseResults
